packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invariance-sdk"
version = "1.0.0"
description = "Python SDK for recording, signing, and verifying AI agent actions against the Invariance backend."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Invariance" }]
keywords = ["ai", "agents", "observability", "receipts", "ed25519"]

dependencies = [
    "httpx>=0.27",
    "pynacl>=1.5",
    "cryptography>=42.0",
    "python-ulid>=2.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.1"]
crewai = ["crewai>=0.1"]
autogen = ["autogen-agentchat>=0.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["invariance"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
